fix(recording): fix macOS system audio sample_rate type and backfill #18 E1 acceptance
- macos.rs: with_sample_rate expects impl Into<i32>; cast spec.sample_rate
  (u32) to i32 to fix E0277 that Windows-host cargo check could not catch
  (cfg(target_os = "macos") branch is never compiled there).
- vendor build.rs: set Swift deployment target to 15.0 via -Xswiftc -target
  so apple-metal's @available(macOS 15.0) APIs don't emit swiftCompatibility56
  back-deployment symbols (unresolvable with CLT-only, no full Xcode).
- acceptance plan: backfill #18 E1 native compile + system audio runtime
  evidence (F-01/F-02/F-06/F-07/F-08 product-level, P-03, C-02). F-03/F-04/
  F-05 remain Blocked; E2/E3/packaging remain Planned.

+++ app/src-tauri/src/audio_capture/macos.rs
@@ -459,7 +459,7 @@ mod platform {
         SCStreamConfiguration::new()
             .with_captures_audio(spec.captures_audio)
             .with_excludes_current_process_audio(spec.excludes_current_process_audio)
-            .with_sample_rate(spec.sample_rate)
+            .with_sample_rate(spec.sample_rate as i32)
             .with_channel_count(spec.channel_count)
     }
